Move AggregatedMetricCard component and story to Shared dirs
